requirements.txt
Django>=3.0,<4.0
djangorestframework==3.11.2
